added moves for pawn
A `Pawn` is a `Piece`.
`Piece` is a struct that is initializable as a chess piece of a specific kind. Eg := Pawn, king, etc.
A `Piece` can determine it's own next possible moves.
To keep the possible moves detemination logic seperate from one another a `PieceNextMovePossibilitiesFinder` interface is defined. Every `Kind` of `Piece` can implement it's own logic on how to determine next possible moves.